Creacion de un MaterialAlertDialog

+++ app/src/main/java/com/fgardila/stores/MainActivity.kt
@@ -1,9 +1,12 @@
 package com.fgardila.stores
 
+import android.content.DialogInterface
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.widget.Toast
 import androidx.recyclerview.widget.GridLayoutManager
 import com.fgardila.stores.databinding.ActivityMainBinding
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import org.jetbrains.anko.doAsync
 import org.jetbrains.anko.uiThread
 
@@ -81,14 +84,34 @@ class MainActivity : AppCompatActivity(), OnClickListener, MainAux {
     }
 
     override fun onDeleteStore(storeEntity: StoreEntity) {
-        doAsync {
-            StoreApplication.database.storeDao().deleteStore(storeEntity)
-            uiThread {
-                mAdapter.delete(storeEntity)
+        val items = arrayOf("Eliminar", "Llamar", "Ir al sitio web")
+
+        MaterialAlertDialogBuilder(this)
+            .setTitle(R.string.dialog_options_title)
+            .setItems(items) { dialog, which ->
+                when(which) {
+                    0 -> confirmDelete(storeEntity)
+                    1 -> Toast.makeText(this, "Llamar", Toast.LENGTH_SHORT).show()
+                    2 -> Toast.makeText(this, "Sitio web", Toast.LENGTH_SHORT).show()
+                }
             }
-        }
+            .show()
     }
 
+    private fun confirmDelete(storeEntity: StoreEntity) {
+        MaterialAlertDialogBuilder(this)
+            .setTitle(R.string.dialog_delete_title)
+            .setPositiveButton(R.string.dialog_delete_confirm) { dialog, i ->
+                doAsync {
+                    StoreApplication.database.storeDao().deleteStore(storeEntity)
+                    uiThread {
+                        mAdapter.delete(storeEntity)
+                    }
+                }
+            }
+            .setNegativeButton(R.string.dialog_delete_cancel, null)
+            .show()
+    }
     /**
      * MainAux
      */
